+++ src/main/java/com/jasmi/xss_scanner/repositories/ScanResultRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ScanResultRepository extends JpaRepository<ScanResult, Integer> {
+public interface ScanResultRepository extends JpaRepository<ScanResult, Long> {
 }
